backend/requirements.txt
# E Lost & Found - Backend dependencies
# Python 3.10+ recommended; use these versions for Python 3.13 (pre-built wheels, no Rust needed)

fastapi>=0.109.2,<0.116
uvicorn[standard]>=0.27.1
pymongo>=4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
boto3>=1.34.34
apscheduler>=3.10.4
pydantic>=2.10.0,<3
pydantic-settings>=2.6.0
email-validator>=2.1.0
